backend/requirements.txt
fastapi[all]==0.110.0
uvicorn==0.29.0
gunicorn==21.2.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
fastapi-limiter==0.1.6
redis==5.0.4
prometheus-fastapi-instrumentator==7.0.0
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.5
